requirements.txt
spi
flask
pillow
rpio
